+++ lib/apis/turbo_firestore_delete_api.dart
@@ -113,8 +113,8 @@ extension TurboFirestoreDeleteApi<T> on TurboFirestoreApi<T> {
           message: 'WriteBatch was null! Deleting without batch..',
           sensitiveData: null,
         );
-        final documentReference =
-            getDocRefById(id: id, collectionPathOverride: collectionPathOverride);
+        final documentReference = getDocRefById(
+            id: id, collectionPathOverride: collectionPathOverride);
         if (transaction == null) {
           _log.debug(
             message: 'Deleting data with documentReference.delete..',
@@ -183,7 +183,8 @@ extension TurboFirestoreDeleteApi<T> on TurboFirestoreApi<T> {
   /// See also:
   /// [deleteDoc] single document deletion
   /// [updateDocs] batch updates
-  Future<TurboResponse<WriteBatchWithReference<Map<String, dynamic>>>> deleteDocs({
+  Future<TurboResponse<WriteBatchWithReference<Map<String, dynamic>>>>
+      deleteDocs({
     required String id,
     WriteBatch? writeBatch,
     String? collectionPathOverride,
@@ -212,7 +213,8 @@ extension TurboFirestoreDeleteApi<T> on TurboFirestoreApi<T> {
       );
       nullSafeWriteBatch.delete(documentReference);
       _log.info(
-        message: 'Adding delete to batch done! Returning WriteBatchWithReference..',
+        message:
+            'Adding delete to batch done! Returning WriteBatchWithReference..',
         sensitiveData: null,
       );
       return TurboResponse.success(