+++ lib/services/turbo_collection_service.dart
@@ -54,7 +54,8 @@ part 'before_after_sync_turbo_collection_service.dart';
 /// - Error handling and logging
 /// - User authentication state synchronization
 abstract class TurboCollectionService<T extends TurboWriteableId<String>,
-    API extends TurboFirestoreApi<T>> extends TurboAuthSyncService<List<T>> with Loglytics {
+        API extends TurboFirestoreApi<T>> extends TurboAuthSyncService<List<T>>
+    with Loglytics {
   /// Creates a new [TurboCollectionService] instance.
   ///
   /// Parameters:
@@ -84,10 +85,6 @@ abstract class TurboCollectionService<T extends TurboWriteableId<String>,
 
   // 👂 LISTENERS ----------------------------------------------------------------------------- \\
   // ⚡️ OVERRIDES ----------------------------------------------------------------------------- \\
-
-  /// Stream of all documents for a given user.
-  @override
-  Stream<List<T>> Function(User user) get stream => (_) => api.streamAllWithConverter();
 
   /// Handles data updates from the Firestore stream.
   ///
@@ -619,7 +616,9 @@ abstract class TurboCollectionService<T extends TurboWriteableId<String>,
       );
       final future = api.createDoc(
         merge: true,
-        writeable: doc.isLocalDefault ? doc : remoteUpdateRequestBuilder?.call(doc) ?? doc,
+        writeable: doc.isLocalDefault
+            ? doc
+            : remoteUpdateRequestBuilder?.call(doc) ?? doc,
         id: doc.id,
         transaction: transaction,
       );
@@ -668,7 +667,9 @@ abstract class TurboCollectionService<T extends TurboWriteableId<String>,
           (await api.createDoc(
             id: doc.id,
             transaction: transaction,
-            writeable: doc.isLocalDefault ? doc : remoteUpdateRequestBuilder?.call(doc) ?? doc,
+            writeable: doc.isLocalDefault
+                ? doc
+                : remoteUpdateRequestBuilder?.call(doc) ?? doc,
             merge: true,
           ))
               .tryThrowFail();
@@ -680,7 +681,9 @@ abstract class TurboCollectionService<T extends TurboWriteableId<String>,
           await api.createDocs(
             id: doc.id,
             writeBatch: batch,
-            writeable: doc.isLocalDefault ? doc : remoteUpdateRequestBuilder?.call(doc) ?? doc,
+            writeable: doc.isLocalDefault
+                ? doc
+                : remoteUpdateRequestBuilder?.call(doc) ?? doc,
             merge: true,
           );
         }